requirements.txt
functions-framework==3.*
requests
pandas
google-cloud-storage
google-cloud-bigquery
flask